requirements.txt
Django==3.2.15
django-debug-toolbar==3.6.0
psycopg2==2.9.3
python-dotenv==0.20.0
requests==2.28.1